Add staging info to rprompt
When playing around with this, it turned out that a solid red hexagon
was too jarring for the "dirty" state (where there are changes but
nothing is staged). So, I spread it back out such that the solid red is
for merge conflicts (legit errors), empty red is for unstaged (not a
hard error, but you wanna know if the working tree is dirty), empty
green is for any number of staged changes (because any number is
legitimate for your next commit anyway), and solid green is for a clean
working tree. That way, the general day-to-day dev work should see
mostly empty hexagons.