Colorize git ref, including the icon
My terminal's color theme's magenta actually isn't half bad. It's kind
of low contrast with my background, so it's still muted. And then it's
less jarring to have the icon be colorized as well, since otherwise the
white-on-black (in my theme) is too high-contrast and doesn't seem
immediately connected to the branch name.