Copycat definition for euclid::fragment
I intend to use this to start splitting up euclid::optics a bit, but
it'll take a lot of elbow grease. For now, I'm just copy/pasting the
definition of euclid::optics into euclid::fragment. The plan is to have
optics handle the colors and for fragments to handle the format strings.
But I'll have to change those airplane wings mid-flight, so I'm going to
work on it one element at a time.