Factor out euclid::element
Now we can see the plan coming together. The optics/fragments split
yields a pattern that can be encapsulated by euclid::element, while
additionally giving us an easier way to configure things because we
don't have to intermingle two separate format string syntaxes (prompt
expansion for colors vs printf for string interpolation).

The logo code is still a bit gnarly because of the %(?.a.b) stuff, but
we can solve that at the same time we factor out another piece of the
puzzle: passing data. Stay tuned.