Invert zstyle storage
By using the input parameter in the context pattern instead of the style
name, we can leverage zstyle to organize certain settings into distinct
hierarchies. That way, we can easily blanket-apply settings (optics,
fragments) to everything within the same hierarchy.

For example, consider storing the logo icon as a euclid fragment.
Before, `euclid::fragment logo X` would have been doing

  zstyle :euclid:fragments logo X

Then looking up the logo with `euclid::fragment logo` would be like
saying

  zstyle -s :euclid:fragments logo reply

But for the more specialized logos, we'd have to separately set & fetch
logo:vicmd and logo:error by hand - the styles don't exist unless we set
them explicitly. This makes it a pain in the ass to update the logo
across all three (logo, logo:vicmd, logo:error) at once. It would be
nice to just say `euclid::fragments logo XX` and have it set all three.

So by inverting things, this change allows us to bake wildcards into the
context patterns. That way, we can just do a single

  zstyle ':euclid:logo*' fragment X

Then when we query for the fragment on ':euclid:logo:vicmd' or
':euclid:logo:error', the wildcard automatically acts as a catch-all,
*unless* we explicitly set more specific fragments.

So that way the user could either set all three logos at the same time
with

  zstyle ':euclid:logo*' fragment XX

or just a specific type of logo with

  zstyle :euclid:logo:vicmd fragment XXX

With the wildcards baked into the euclid::{optics,fragment} functions,
the user doesn't even have to be particularly aware that this is what's
happening under the hood. Magic. ✨