Rename euclid::element to euclid::fragment
I'm planning on introducing a euclid::elements system similar to
euclid::{optics,data}, and I need to free up the name euclid::element
for other purposes. As it turns out, euclid::fragment will be a semantic
name: elements may be composed of fragments, which are little snippets
of formatted text. (And the Euclid book Elements has several fragments
that were found.)